requirements.txt
asgiref==3.7.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.1
django-bootstrap5==23.4
django-localflavor==4.0
EpicSDKScripts @ file:///home/rbdundas/projects/InsuranceProductivity/wheels/epicsdkscripts-0.1.0-py3-none-any.whl#sha256=f2e19089d6649d19f3e55fd689c071273d8c8d91ee3ab01904ed77d34a3538d1
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.7
lxml==5.2.2
numpy==1.26.4
oauthlib==3.2.2
pandas==2.2.1
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-stdnum==1.19
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
